Rename file name to QueryBoundaryWrapper
Following from the component rename itself.
